+++ ui/__ColorPixmap.py
