+++ ui/__ConnectionSetupWindow.py
@@ -11,7 +11,7 @@ from PySide6.QtWidgets import (
 from PySide6.QtGui import QIcon, QIntValidator
 from PySide6.QtCore import Qt
 
-from ui.ui_utilities import load_and_color_icon_pixmap
+from ui.__ui_utilities import load_and_color_icon_pixmap
 
 
 class ConnectionSetupWindow(QDialog):
