+++ ui/__IconLabel.py
@@ -5,7 +5,7 @@ from PySide6.QtWidgets import (
 )
 from PySide6.QtCore import Qt
 
-from ui.ui_utilities import load_and_color_icon_pixmap
+from ui.__ui_utilities import load_and_color_icon_pixmap
 
 
 class IconLabel(QWidget):
