+++ ui/__ui_utilities.py
@@ -3,7 +3,7 @@ from PySide6.QtGui import QPixmap
 from PySide6.QtCore import QSize
 from PySide6.QtGui import QIcon
 
-from ui.ColorPixmap import ColorPixmap
+from ui.__ColorPixmap import ColorPixmap
 
 
 def image_to_qpixmap(path):
